requirements.txt
apache-airflow
kafka-python
cassandra-driver
pyspark
spark